timeline logic split and style
