+++ project.todo
@@ -2,6 +2,14 @@ Basics:
   ✔ branch - feature @done (2/21/2019, 8:42:17 AM)
   ✔ dev branch @done (2/21/2019, 8:42:19 AM)
   ✔ bugfixes @done (2/21/2019, 8:42:20 AM)
+
+
+STORY DRIVEN DEVELOPMET PRINICIPLES: 
+  ☐ Storybook - key
+  ☐ Juniors work on non-connected components
+  ☐ Seniors containers
+  ☐ components requirements as a withstory element. 
+
 Aim:
   ☐ to create a responsive blog like for #100daysofcode
   ☐ or bachata london 2.0 (admin set events, dynamic timeline of events location etc.)
@@ -28,6 +36,7 @@ Aim:
   
 
 Frontend:
+  ☐ query pagination fetchMore @today
   ☐ read more about branches management (features, bugfixes) @low
   ✔ react @critical @done (2/21/2019, 9:40:50 PM)
   ☐ react-blog @critical
@@ -38,7 +47,7 @@ Frontend:
   ✔ getInitialProps and async data fetch in nextjs @high @done (2/21/2019, 9:41:00 PM)
   ✔ containers / components logic split.  @low @done (2/21/2019, 9:41:01 PM)
   ✔ containers / components export indexing.  @low @done (2/21/2019, 9:41:02 PM)ß
-  ☐ nextjs json config imports
+  ✔ nextjs json config imports @done (3/19/2019, 3:16:07 PM)
   ☐ nextjs translations / i[]
   ☐ nextjs SEO best practises (https://github.com/zeit/next.js/tree/canary/examples/with-next-seo)
   ☐ experiment with the apollo client data request from get InitialProps / within pages only
@@ -49,8 +58,8 @@ Frontend:
   ☐ learn redux-thunk
   
 Backend:
-  ☐ custom ssr with nextjs @low
-  ☐ static root files (with express) @low
+  ✔ custom ssr with nextjs @low @done (3/19/2019, 3:16:44 PM)ß
+  ✔ static root files (with express) @low @done (3/19/2019, 3:16:47 PM)
   ☐ mongodb @high
   ☐ keystonejs
   ✔ apollo @critical (https://github.com/zeit/next.js/tree/canary/examples/with-apollo-and-redux) @done (2/26/2019, 6:00:07 PM)
